Got rid of old code and renamed AutoRed.java to Autonomous.java
